hireready-backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
google-generativeai>=0.3.0
firebase-admin>=6.2.0
pdfplumber>=0.10.0
slowapi>=0.1.9
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
